C++: Split off the createLSParser tests into their own file.
