+++ cpp/ql/test/query-tests/Security/CWE/CWE-611/tests.h
@@ -2,3 +2,13 @@
 
 class SecurityManager;
 class InputSource;
+
+class AbstractDOMParser {
+public:
+	AbstractDOMParser();
+
+	void setDisableDefaultEntityResolution(bool); // default is false
+	void setCreateEntityReferenceNodes(bool); // default is true
+	void setSecurityManager(SecurityManager *const manager);
+	void parse(const InputSource &data);
+};
